[REFACTOR] 파일을 나누고 패키지 구조를 변경했어요.
