+++ presentation/src/main/java/com/nexters/knownknowns/presentation/navigation/NavController.kt
@@ -1,0 +1,5 @@
+package com.nexters.knownknowns.presentation.navigation
+
+class NavController {
+
+}
