+++ presentation/src/main/java/com/nexters/knownknowns/presentation/navigation/Screen.kt
@@ -1,0 +1,9 @@
+package com.nexters.knownknowns.presentation.navigation
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed interface Screen : NavKey {
+    data object Main : Screen
+}
